feat(backend,docs): implement portfolio CMS backend and update project documentation

+++ backend/src/modules/uploads/uploads.controller.ts
@@ -1,10 +1,13 @@
 import {
   Controller,
   Delete,
+  ForbiddenException,
   Param,
   ParseIntPipe,
   Post,
+  Req,
   UploadedFile,
+  UseGuards,
   UseInterceptors,
 } from '@nestjs/common';
 import { FileInterceptor } from '@nestjs/platform-express';
@@ -12,7 +15,12 @@ import { diskStorage } from 'multer';
 import { extname, join } from 'path';
 import { ImagesService } from '../images/images.service';
 import { existsSync, mkdirSync } from 'fs';
+import { JwtAuthGuard } from '../auth/guards/jwt-auth.guard';
+import { UserRoles, type AuthenticatedRequest } from 'src/utils/types';
+import { ActiveUserGuard } from '../auth/guards/active-user.guard';
+import { UserOwnershipGuard } from '../auth/guards/user-ownership.guard';
 
+@UseGuards(JwtAuthGuard, ActiveUserGuard, UserOwnershipGuard)
 @Controller('upload')
 export class UploadsController {
   constructor(private imagesService: ImagesService) {}
@@ -53,7 +61,14 @@ export class UploadsController {
   async uploadImage(
     @Param('userId', ParseIntPipe) userId: number,
     @UploadedFile() file: Express.Multer.File,
+    @Req() req: AuthenticatedRequest,
   ) {
+    const { sub, role } = req.user;
+    if (Number(sub) !== userId || Number(role) !== UserRoles.SYSADMIN) {
+      throw new ForbiddenException(
+        'You can only upload images to your own account.',
+      );
+    }
     // Salva registro no banco
     const saved = await this.imagesService.saveImage({
       f_userId: userId,
@@ -67,7 +82,22 @@ export class UploadsController {
   }
 
   @Delete(':imageId')
-  async deleteImage(@Param('imageId', ParseIntPipe) imageId: number) {
+  async deleteImage(
+    @Param('imageId', ParseIntPipe) imageId: number,
+    @Req() req: AuthenticatedRequest,
+  ) {
+    const { sub, role } = req.user;
+    const image = await this.imagesService.findOne(imageId);
+    if (!image) {
+      throw new ForbiddenException('Image not found.');
+    }
+
+    if (Number(sub) !== image.f_userId || Number(role) !== UserRoles.SYSADMIN) {
+      throw new ForbiddenException(
+        'You can only delete images to your own account.',
+      );
+    }
+
     return this.imagesService.delete(imageId);
   }
 }
